BUG Bug com criação de ficha

+++ public/service/Back/reqFicha.js
@@ -35,7 +35,7 @@ class FichaServices {
         )
       })
       const result = await response.json();
-       if (result.message === "Não autorizado") {
+       if (!result && result.message === "Não autorizado") {
             alert("Sessão expirada faça login novamente!")
             localStorage.clear()
             window.location.href = "/"
